MOZ-392 Fix bug in 3-tables version

+++ src/main/java/com/mapofzones/endpointchecker/domain/NodeLcdAddress.java
@@ -43,7 +43,7 @@ public class NodeLcdAddress {
     private Boolean isAlive;
 
     @Column(name = "RESPONSE_TIME")
-    private Long responseTime;
+    private Integer responseTime;
 
     @Column(name = "LAST_ACTIVE")
     private LocalDateTime lastActive;
